FREventMessages and FREventMessage moved to common project
