+++ pom.xml
@@ -50,7 +50,7 @@
     </parent>
 
     <properties>
-        <uk.bom.version>2.0.3</uk.bom.version>
+        <uk.bom.version>2.0.4-SNAPSHOT</uk.bom.version>
         <consent.api.version>2.0.2</consent.api.version>
     </properties>
