+++ secure-api-gateway-ob-uk-rs-resource-store/secure-api-gateway-ob-uk-rs-resource-store-api/src/main/java/com/forgerock/sapi/gateway/rs/resource/store/api/admin/events/FRDataEventsConverter.java
@@ -18,7 +18,7 @@ package com.forgerock.sapi.gateway.rs.resource.store.api.admin.events;
 import java.util.Objects;
 
 import com.forgerock.sapi.gateway.ob.uk.common.datamodel.converter.mapper.FRModelMapper;
-import com.forgerock.sapi.gateway.rs.resource.store.datamodel.events.FREventMessage;
+import com.forgerock.sapi.gateway.ob.uk.common.datamodel.event.FREventMessage;
 import com.forgerock.sapi.gateway.rs.resource.store.repo.entity.event.FREventMessageEntity;
 
 import uk.org.openbanking.datamodel.event.OBEventNotification1;
