--- secure-api-gateway-ob-uk-rs-resource-store/secure-api-gateway-ob-uk-rs-resource-store-datamodel/src/main/java/com/forgerock/sapi/gateway/rs/resource/store/datamodel/events/FREventMessages.java
@@ -1,56 +1,0 @@
-/*
- * Copyright © 2020-2022 ForgeRock AS (obst@forgerock.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.forgerock.sapi.gateway.rs.resource.store.datamodel.events;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-import uk.org.openbanking.datamodel.event.OBEventNotification1;
-
-/**
- * A list of events, used to import and update events through events admin api 'DataEventsApi'
- */
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@Builder
-public class FREventMessages {
-    @JsonProperty("client_id")
-    private String apiClientId;
-    @JsonProperty("events")
-    private List<OBEventNotification1> obEventNotification1List = new ArrayList();
-
-
-    public FREventMessages obEventNotification1List(List<OBEventNotification1> obEventNotification2List) {
-        this.obEventNotification1List = obEventNotification2List;
-        return this;
-    }
-
-    public FREventMessages addOBEventNotification1Item(OBEventNotification1 obEventNotification2) {
-        if(Objects.isNull(obEventNotification1List)) {
-            obEventNotification1List = new ArrayList<>();
-        }
-        this.obEventNotification1List.add(obEventNotification2);
-        return this;
-    }
-}
